+++ _build/data/transport.system_settings.php
@@ -9,4 +9,11 @@ return [
         'namespace' => 'twig',
         'area' => 'general',
     ],
+    [
+        'key' => 'twig.document_pass',
+        'value' => '1',
+        'xtype' => 'combo-boolean',
+        'namespace' => 'twig',
+        'area' => 'general',
+    ],
 ];